fix date loading between pages
