+++ frontend/src/util/search.ts
@@ -1,4 +1,8 @@
 import { LazyQueryExecFunction, QueryResult } from '@apollo/client'
+import { QueryBuilderState } from '@components/QueryBuilder/QueryBuilder'
+import { resetRelativeDates } from '@highlight-run/ui'
+import { useErrorSearchContext } from '@pages/Errors/ErrorSearchContext/ErrorSearchContext'
+import { useSearchContext } from '@pages/Sessions/SearchContext/SearchContext'
 import log from '@util/log'
 import { useEffect, useRef, useState } from 'react'
 
@@ -16,6 +20,12 @@ export function usePollQuery<T, U>({
 }) {
 	const pollTimeout = useRef<number>()
 	const [numMore, setNumMore] = useState<number>(0)
+
+	const { searchQuery, setSearchQuery } = useSearchContext()
+	const {
+		searchQuery: errorSearchQuery,
+		setSearchQuery: setErrorSearchQuery,
+	} = useErrorSearchContext()
 
 	useEffect(() => {
 		// setup a polling interval for sessions after the current date range
@@ -53,9 +63,34 @@ export function usePollQuery<T, U>({
 	return {
 		numMore,
 		reset: () => {
+			resetRelativeDates()
 			clearTimeout(pollTimeout.current)
 			pollTimeout.current = undefined
 			setNumMore(0)
+
+			const currentState = JSON.parse(searchQuery) as QueryBuilderState
+			const newRules = currentState.rules.filter(
+				(rule) => rule[0] !== 'custom_created_at',
+			)
+			setSearchQuery(
+				JSON.stringify({
+					isAnd: currentState.isAnd,
+					rules: newRules,
+				}),
+			)
+
+			const currentErrorState = JSON.parse(
+				errorSearchQuery,
+			) as QueryBuilderState
+			const newErrorRules = currentErrorState.rules.filter(
+				(rule) => rule[0] !== 'error-field_timestamp',
+			)
+			setErrorSearchQuery(
+				JSON.stringify({
+					isAnd: currentState.isAnd,
+					rules: newErrorRules,
+				}),
+			)
 		},
 	}
 }
